Fake is_admin() to false when BP pushState's AJAX call is running.
Since is_admin() always returns true when AJAX is running, some
trickery is needed to fake is_admin() to true.  This will fix issues with
plugins doing conditional checks on is_admin() to load their frontend
code on our AJAX request.

An example of this is with the GD bbPress Attachments plugin.  This commit
now removes the previous hack of manually loading up GD since this is no
longer necessary with the is_admin() override.